requirements/base.txt
celery==5.2.2
Django==3.2.13
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-filter==21.1
django-simple-history==3.0.0
djangorestframework==3.13.1
drf-spectacular[sidecar]==0.20.2
factory-boy==3.2.0
Pillow==9.0.1
psycopg2-binary==2.9.1
python-json-logger==2.0.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
weasyprint==54.1
